pci: PCI-E driver for Baikal-M with SDK-M 5.5 firmware
This driver works with firmware from SDK-M 5.5. dw-pcie should be
used with older firmware (such as SDK-M 5.3)

Note: this driver uses `baikal,bm1000-pcie` as a compatible string.
Hence the kernel with this driver can run on Baikal-M boards with
older firmware (and use dw-pcie driver instead of this one).

X-feature-Baikal-M
X-DONTUPSTREAM